Update Verifying-ERC20.md
In line 14, is this a direct quote? I added an "and" as a correction, but if this is a direct quote, then leave it as you wrote it.
In line 87, "since it must *quantify* over the set of all accounts", Is over normally used after quantify in these types of documents? I feel the 'over' is unnecessary, but if it's normally used, then keep it. There's another instance in line 193.
In line 216, is "The problem is due to a bug in `mint`" an example of a type of error verisol would report? I'm trying to figure out the right punctuation.
In line 218, I changed 'between' to 'among'. Between is for 2 things, and among is for 3 or more things. I'm not sure which one is correct for you, so you can choose depending on how many variables you're talking about.
In line 246, is "instruments" being used as a verb here "The tool instruments the contract to find"?

Not sure why, but in some lines or words, it crossed out the entire thing and rewrote it, even if I just changed one thing.